migration DB

+++ src/main/java/com/dataart/cerebro/exception/NotFoundException.java
@@ -2,7 +2,7 @@ package com.dataart.cerebro.exception;
 
 public class NotFoundException extends RuntimeException{
 
-    public NotFoundException(String object, int id) {
+    public NotFoundException(String object, long id) {
         super(String.format("%s with ID (%d) doesn't exist", object, id));
     }
 }
